odontocare/services/citas-service/requirements.txt
requests
responses
pybikes
pandas
matplotlib
pytest
Flask
Jinja2
scipy
numpy
pillow
pymongo
flask-sqlalchemy
PyJWT
jsonschema
werkzeug
marshmallow
